backend/internal/middleware/jwt_middleware.go
package middleware

import (
	"net/http"
	"os"

	jwtmiddleware "github.com/auth0/go-jwt-middleware"
	"github.com/form3tech-oss/jwt-go"
)

func NewJWTMiddleware() *jwtmiddleware.JWTMiddleware {
	return jwtmiddleware.New(jwtmiddleware.Options{
		ValidationKeyGetter: func(token *jwt.Token) (interface{}, error) {
			return []byte(os.Getenv("AUTH0_CLIENT_SECRET")), nil
		},
		SigningMethod: jwt.SigningMethodHS256,
		Extractor:     jwtmiddleware.FromAuthHeader,
	})
}

func JwtVerify(jwtMiddleware *jwtmiddleware.JWTMiddleware) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			err := jwtMiddleware.CheckJWT(w, r)
			if err != nil {
				http.Error(w, "Forbidden", http.StatusForbidden)
				return
			}
			next.ServeHTTP(w, r)
		})
	}
}
